chore: move back client types
