+++ packages/vite/types/importMeta.d.ts
@@ -1,13 +1,25 @@
-import type {
-  ImportMeta as ViteImportMeta,
-  ImportMetaEnv as ViteImportMetaEnv
-} from '../client/types'
+import type { ViteHotContext } from './hot'
+import type { ImportGlobEagerFunction, ImportGlobFunction } from './importGlob'
 
-declare global {
-  interface GlobOptions {
-    as?: string
-  }
+export interface ImportMetaEnv {
+  [key: string]: any
+  BASE_URL: string
+  MODE: string
+  DEV: boolean
+  PROD: boolean
+  SSR: boolean
+}
 
-  interface ImportMeta extends ViteImportMeta {}
-  interface ImportMetaEnv extends ViteImportMetaEnv {}
+export interface ImportMeta {
+  url: string
+
+  readonly hot?: ViteHotContext
+
+  readonly env: ImportMetaEnv
+
+  glob: ImportGlobFunction
+  /**
+   * @deprecated Use `import.meta.glob('*', { eager: true })` instead
+   */
+  globEager: ImportGlobEagerFunction
 }
